Add role update functionality; implement endpoint for updating account roles and enhance service logic

+++ src/main/java/com/gym/dao/RoleDao.java
@@ -12,5 +12,5 @@ public interface RoleDao {
 
     List<Role> getAllRoles();
 
-    void updateRolesByAccountId(int accountId, List<Long> roleIds);
+    void updateRolesByAccountId(long accountId, List<Long> roleIds);
 }
